+++ src/components/pokemon/PokemonAttribute.tsx
