+++ src/components/pokemon/PokemonAttributes.tsx
